requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1
django-extensions==3.0.9
djangorestframework==3.11.1
djangorestframework-yaml==2.0.0
idna==2.10
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.2.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
traitlets==5.0.4
transitions==0.7.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==0.16.0
wtfpython==0.2.1
